+++ doc/source/CHANGELOG.rst
@@ -32,6 +32,7 @@ your Python installation to at least version 3.5.
 - estimators: n_jobs setting does not look for OMP_NUM_THREADS, but for PYEMMA_NJOBS and SLURM_CPUS_ON_NODE to avoid
   multiplying OpenMP threads with PyEMMA processes. On SLURM the number of allocated cores is used.
   If nothing is set, the physical cpu count is considered.
+- msm: calling score_cv does not modify the object anymore. #1178
 
 
 2.4 (05-19-2017)